+++ changelog.md
@@ -3,10 +3,11 @@
   * Allows adding crafting recipes to the anvil via data pack
   * Adds a data pack that makes so iron, gold and diamond equipment must be forged in an anvil, from copper, wood/leather and gold respectively
   * Includes EMI integration
-    * Yes, EMI good, I will support it only
 * Added a new attribute for Exhaustion
   * `insanesurvivaloverhaul:exhaustion_multiplier`: Multiplies the exhaustion applied to the player. Defaults to 1.
   * Effective hunger now uses this attribute
+* Added back EMI integration
+  * Yes, EMI good, I will support it only2
 * Fixed pouch tooltip being too high
 
 ## 5.4.2.2